+++ packages/drizzledb-pg/src/schema/index.ts
@@ -1,24 +1,1 @@
-import { pgSchema, text, timestamp, uuid } from 'drizzle-orm/pg-core'
-
-import { citext } from './custom-types.js'
-
-export { count } from 'drizzle-orm'
-
-export const authSchema = pgSchema('auth')
-
-export const users = authSchema.table('users', {
-  id: uuid('id').primaryKey().defaultRandom(),
-  username: citext('username').notNull().unique(),
-  hashedPassword: text('hashed_password').notNull(),
-})
-
-export const sessions = authSchema.table('sessions', {
-  id: text('id').primaryKey(),
-  userId: uuid('user_id')
-    .notNull()
-    .references(() => users.id),
-  expiresAt: timestamp('expires_at', {
-    withTimezone: true,
-    mode: 'date',
-  }).notNull(),
-})
+export * from './auth-schema.js'
