+++ packages/lucia-auth/src/repository/index.ts
@@ -1,19 +1,32 @@
 import { Argon2id } from 'oslo/password'
 
 import { db } from '@packages/drizzledb-pg/client'
-import * as schema from '@packages/drizzledb-pg/schema'
 import { users } from '@packages/drizzledb-pg/schema'
 
-export type NewUser = Omit<typeof schema.users.$inferInsert, 'id' | 'hashedPassword'> & {
+export type NewUser = Omit<typeof users.$inferInsert, 'id' | 'hashedPassword'> & {
   password: string
 }
 
-export type User = typeof schema.users.$inferSelect
+export type User = typeof users.$inferSelect
+
+export const findByUsername = async (username: string): Promise<User | undefined> => {
+  return db.query.users.findFirst({
+    columns: { id: true, username: true, hashedPassword: true },
+    where: (col, { eq }) => eq(col.username, username),
+  })
+}
 
 export const createUser = async (user: NewUser): Promise<string | undefined> => {
   const hashedPassword = await new Argon2id().hash(user.password)
+
+  await db.query.users.findMany({
+    with: {
+      sessions: true,
+    },
+  })
+
   return db
-    .insert(schema.users)
+    .insert(users)
     .values({ username: user.username, hashedPassword })
     .returning({ id: users.id })
     .then((res) => res[0]?.id)
